requirements.txt
# Core and Integrations
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langchain-huggingface
langchain-ollama
langchain-chroma
langchain-nvidia-ai-endpoints

# Embeddings
sentence-transformers

# Web Application
streamlit
python-dotenv

# Database
sqlalchemy
pymysql

# Document Processing
unstructured

# Output Formatting
fpdf
img2pdf
markdown

# Visualization
plotly
kaleido
matplotlib